Move remaining tests from tests/ to src/.
