+++ src/python/pants/backend/project_info/list_integration_test.py
