+++ src/python/pants/engine/goal_integration_test.py
