+++ src/python/pants/init/BUILD
@@ -17,7 +17,12 @@ python_sources(
 python_tests(
     name="tests",
     overrides={
-        "load_backends_integration_test.py": {"timeout": 900},
+        "load_backends_integration_test.py": {
+            "dependencies": [
+                "testprojects:pants_plugins_directory",
+            ],
+            "timeout": 900,
+        },
         "logging_test.py": {"dependencies": ["//BUILD_ROOT:files"]},
         ("options_initializer_test.py", "plugin_resolver_test.py"): {
             "dependencies": [
